requirements.txt
click
flask
Pillow
requests_html
pytest
vcrpy
